Add files via upload
The rules of the game: You need to move the tiles and every time you move one, another tile pops up in a random location. When two tiles with the same number on them collide with one another as you move them, they will merge into one tile with the sum of the numbers. "W", "S", "A", "D" to move up, down, left, and right respectively.