Add files via upload
The game, mastermind, is about player A trying to guess the player B's hidden color combination, with a maximum of 10 tries, each time getting a feedback from player B. The feedback is as follows: Black = Peg is the right color, in the right location. White = Peg is the right color, but in the wrong location. The game is played with 4 pegs and 6 colors, which gives a total of 1296 different patterns. Using several algorithms to solve the game, I am able to find the exact combination in a maximum of 6 tries. In addition, the algorithms extend to a greater number of pegs and colors, however, implying that the number of tries will exceed 6 tries.